+++ transformers_lightning/datasets/__init__.py
@@ -1,6 +1,6 @@
 # triples of name as key, yaml config file and data file as values
 from transformers_lightning.datasets.compressed_dataset import CompressedDataset  # noqa: F401
 from transformers_lightning.datasets.iterable_dataset import TransformersIterableDataset  # noqa: F401
-from transformers_lightning.datasets.map_dataset import MapDataset  # noqa: F401
+from transformers_lightning.datasets.map_dataset import TransformersMapDataset  # noqa: F401
 from transformers_lightning.datasets.stack_dataset import StackDataset  # noqa: F401
 from transformers_lightning.datasets.super_dataset import SuperDataset  # noqa: F401